Rename module `overlay.py` to `embed.py`
